requirements.txt
jsonschema==3.2.0
packaging==24.0
python-dateutil==2.9.0.post0
